some styles added on Signup and Manage conversation with Translation
