+++ client/src/App.js
@@ -1,7 +1,7 @@
 import React from "react";
-import Signup from "./Signup.js";
-import Signin from "./Signin.js";
-import Message from "./Message.js";
+import Signup from "./components/Signup/Signup.js";
+import Signin from "./components/Signin/Signin";
+import Message from "./components/Message/Message.js";
 import Home from "./components/Home";
 import { BrowserRouter as Router, Route } from "react-router-dom";
 
@@ -13,7 +13,7 @@ function App() {
       <div>
         {/* <NavTabs /> */}
         <Route exact path="/" component={Home} />
-        <Route exact path="/SignUp" component={Signup} />
+        <Route exact path="/Signup" component={Signup} />
         <Route exact path="/Signin" component={Signin} />
         <Route exact path="/Message" component={Message} />
       </div>
