+++ client/src/components/Signin/Signin.css
@@ -1,0 +1,11 @@
+#containerSmall {
+  background-color: #bbd5f3;
+  height: 700px;
+  margin-left: 0px;
+}
+
+#containerBig {
+  background-color: #77a4d8;
+  height: 700px;
+  margin-right: 0px;
+}
